python/requirements.txt
pyserial==3.5
PySDL2==0.9.16

